Move test_array_function to directories
